+++ Pinta.Tools/Tools/ColorPickerTool.cs
@@ -63,7 +63,10 @@ public class ColorPickerTool : BaseTool
 	private bool SampleLayerOnly => SampleTypeDropDown.SelectedItem.GetTagOrDefault (false);
 
 	public void UseImageSampling ()
-		=> SampleTypeDropDown.SelectedIndex = 1;
+	{
+		SampleSizeDropDown.SelectedIndex = 0;
+		SampleTypeDropDown.SelectedIndex = 1;
+	}
 
 	public override Gdk.Cursor DefaultCursor {
 		get {